+++ clients/android/trace/app/src/main/java/com/example/brandon/trace/FileListItem.java
@@ -6,8 +6,8 @@ package com.example.brandon.trace;
 public class FileListItem {
     public String fileName;
     public String status;
-    public int size;
-    public int progress;
+    public double size;
+    public double progress;
     public boolean showProgress;
 
     public FileListItem(String fileName, String status) {
